Move DictItem to it's own file
